+++ src/main/java/me/themoo/extranpc/util/SupportLinks.java
@@ -1,0 +1,16 @@
+package me.themoo.extranpc.util;
+
+/**
+ * Official support links for ExtraNPC.
+ */
+public final class SupportLinks {
+
+    public static final String DISCORD_INVITE = "https://discord.com/invite/fJTSA6vnVQ";
+    public static final String DISCORD_USER = "@m_1z.4";
+    public static final String SPIGOT_RESOURCE_ID = "138244";
+    public static final String SPIGOT_URL =
+            "https://www.spigotmc.org/resources/extranpc-plugin." + SPIGOT_RESOURCE_ID + "/";
+
+    private SupportLinks() {
+    }
+}